Dspic
Squash merge of all things required for getting the DSPIC33 that we use on logger and sensor working


* First stab at DSPIC33EPxxxGP CAN driver

Two things I don't like about it:
1. Can only transmit one message at a time
2. Reception of CAN messages is handled in an interrupt context,
   should be doable in a polling fashion

Both of these are solveable, but I haven't put the time into them yet.

* Create a dspic CAN driver test case

It runs on logger board. If you want it to run on sensor or something,
you're gonna need to change the LED toggle functions.

The test runs in loopback mode, in fact the whole driver does. This
will change in the next commit.

* Add init argument for whether to run dspic driver in loopback mode

The test doesn't really work unless you're in loopback mode. So this
lets you put it in loopback mode.

* Review Fixes

Changes in this commit:
    - Remove unnecessary comment about "MUST CHANGE BEFORE MERGING"
    - Add comment saying that this is a 16-bit part, and left
    shifting an unsigned by 8 is not an issue
    - Change compile guard name to reflect part name
    - Correct wrong part name in comment in header file