Create message_types file defining all the SIDs
Using #defines for SIDs because screw good practice (while it usually
is better practice to use const ints for global constants like this
because of better debuggability, we use defines since it saves having
to access memory whenever reading the value of these
constants. Usually your compiler would optimize out those memory
accesses, but all of the busproj projects compile at optimization
level 0 for reasons, so I don't know that it does that).

Some information about how message types are structured is not in this
file, it is only available in the message table spreadsheet (info like
message lengths and data formats). The spreadsheet is linked so that
data is viewable, the data is not duplicated because I worry I'd copy
something over wrong and then we no longer have a canonical version of
truth.