Message types (#3)
* Create message_types file defining all the SIDs

Using #defines for SIDs because screw good practice (while it usually
is better practice to use const ints for global constants like this
because of better debuggability, we use defines since it saves having
to access memory whenever reading the value of these
constants. Usually your compiler would optimize out those memory
accesses, but all of the busproj projects compile at optimization
level 0 for reasons, so I don't know that it does that).

Some information about how message types are structured is not in this
file, it is only available in the message table spreadsheet (info like
message lengths and data formats). The spreadsheet is linked so that
data is viewable, the data is not duplicated because I worry I'd copy
something over wrong and then we no longer have a canonical version of
truth.

* Create a build_can_message function

* Write some unit tests against build_can_message

Unit tests added in tests/unit. There's a Makefile. Yes it's
shitty. Feel free to fix it if it offends you. It works though, so I'm
happy-ish with it.

* Write some utility functions and associated unit tests

Wrote functions to determine message type of message, as well as
whether a message is a sensor message and whether a message is a debug
message

* Write a build_printf_can_message function

The header file doesn't have a good comment on it yet, because the
function signature is convoluted for weird C reasons, and because it's
past 1 in the morning and I really need to go to bed.

There's also a single unit test, so I guess it's probably fine, right?

* Create DEBUG macro, and write an associated unit test

* Review Changes:

changes in this commit:
    - move CANLIB_VERSION and CANLIB_BIT_TIME_US defines to can.h
    - add some comments to describe tricky sections of the code
    - add spaces after if's because Alex complained
    - indentation changes
    - change function name from debug_level_message to message_debug_level

* Whitespace changes

I have run all the new files through my astyle config. Most of the changes
were to the opening brace of functions, since K&R style says they should be
on a different line. Also fixed all of the if( lack of spaces

* Remove misleading comment