Define board IDs for all existing boards
We currently have 6 boards, and I'm leaving IDs for spare versions of
each one. Board names are sorted alphabetically, and IDs are allocated
starting at 1.

The idea isn't that you'll use these defines for anything other than
reference. The definition BOARD_UNIQUE_ID should be setup in the MPLAB
configuration xml file, so that all files are compiled with it, not just
those that are including message_types.h. The reason these numbers are
being commited is just to have a single source of truth for the whole
project.