Create buffering layer for CAN message receiving
We have an issue in several of the drivers that the CAN message received
callback executes in an ISR context, but we don't want to force end
users to write interrupt style (efficient, thread-safe, etc) code to
handle CAN messages.

This buffering code (found in util) provides a function that you can use
as the callback which pushes the received CAN message into a ring buffer
that can be read out of from application code at some later time. The
ring buffer library needs to be initialized with a memory pool by the
application code, because it isn't our place to say how much memory
you're supposed to allocate to this library.

This commit also creates a unit test file
(tests/unit/can_buffering_layer.c) that makes sure that this module
works, at least on my x86 laptop. I have not tested this module on a
microcontroller.