Run buffering layer unit tests on radio board
One unit test was failing, that has been fixed. The test tried to buffer
10 CAN messages, and then ensure that there was no room left in the
buffering layer. The buffer was set to be 140 bytes long, because on my
laptop every buffer element took up 14 bytes. Due to alignment
differences, a buffer element on the microcontroller takes up only 12
bytes, so we could fit 11 buffer elements in 140 bytes. Solution was to
remove the check that the buffer was full - it doesn't actually do
anything.

So anyways, the test passes. We're good.