Add all other unit tests to on-hardware test project
This required only minor changes to the unit tests, none functional. The
changes all had to do with not calling printf on the hardware, since it
doesn't have a stdout. This is done with the NO_PRINTF macro that's
defined in the project, which silences all print statements.

All 4 unit tests pass btw. So at least we know how to code a bit.