Add overflow flag to can buffering layer
The flag is set whenever you try to push a message into a full buffer.
You can read the flag with rcvb_has_overflowed, and clear is with
rcvb_clear_overflow_flag.

In an ideal world this buffer will never overflow because the main loop
will deal with messages faster than they come in. Unfortunately, we do
not live in the ideal world, so this just allows you to check for the
error condition.