Create buffering layer for CAN message receiving (#7)
* Create buffering layer for CAN message receiving

We have an issue in several of the drivers that the CAN message received
callback executes in an ISR context, but we don't want to force end
users to write interrupt style (efficient, thread-safe, etc) code to
handle CAN messages.

This buffering code (found in util) provides a function that you can use
as the callback which pushes the received CAN message into a ring buffer
that can be read out of from application code at some later time. The
ring buffer library needs to be initialized with a memory pool by the
application code, because it isn't our place to say how much memory
you're supposed to allocate to this library.

This commit also creates a unit test file
(tests/unit/can_buffering_layer.c) that makes sure that this module
works, at least on my x86 laptop. I have not tested this module on a
microcontroller.

* Create an on-device unit test project for pic18f

The project is targetting the radio board, but that's only for LEDs. It
should run on any pic18f26k83.

I wrote this commit from a Starbucks, I have no idea if the unit tests
actually pass on hardware. If they do pass, LED3 (blue) should light up,
else LED1 (red) will light up.

* Run buffering layer unit tests on radio board

One unit test was failing, that has been fixed. The test tried to buffer
10 CAN messages, and then ensure that there was no room left in the
buffering layer. The buffer was set to be 140 bytes long, because on my
laptop every buffer element took up 14 bytes. Due to alignment
differences, a buffer element on the microcontroller takes up only 12
bytes, so we could fit 11 buffer elements in 140 bytes. Solution was to
remove the check that the buffer was full - it doesn't actually do
anything.

So anyways, the test passes. We're good.

* Add all other unit tests to on-hardware test project

This required only minor changes to the unit tests, none functional. The
changes all had to do with not calling printf on the hardware, since it
doesn't have a stdout. This is done with the NO_PRINTF macro that's
defined in the project, which silences all print statements.

All 4 unit tests pass btw. So at least we know how to code a bit.

* Buffering layer review comments fixed

  - Rename buffering layer files to can_rcv_buffer.c/h
  - Remove stupid function names form can_rcv_buffer. They're all
prefixed with rcvb_ in order to avoid collisions
  - Remove unneeded stdio include in can buffering layer
  - Update unit tests to account for these changes

* Add overflow flag to can buffering layer

The flag is set whenever you try to push a message into a full buffer.
You can read the flag with rcvb_has_overflowed, and clear is with
rcvb_clear_overflow_flag.

In an ideal world this buffer will never overflow because the main loop
will deal with messages faster than they come in. Unfortunately, we do
not live in the ideal world, so this just allows you to check for the
error condition.