Rework can message formatting functions
Replace the generic build_can_msg function with a set of more granular
functions for each message type. This is because the current
implementation of build_can_message exposes too much of the internal
message format (data that is passed to it must already be correctly
formatted). Its original intent was to be used inside wrapper functions.
However, it is a bit too rigid to be used effectively as such, and there
is very little that is common between all message types.

The unit tests have been updated to use the new functions.