Create DEBUG_RADIO_CMD CAN message type
It's effectively the same as a debug printf CAN message, but whenever
radio board receives one of these messages, it should handle each ASCII
byte of data as though it had received it from the XBEE.

The purpose of having such a CAN message is to be able to test/debug the
whole system without having RLCS set up. Radio board will be receiving
radio traffic from RLCS, so if we want to test how it handles that
traffic without using RLCS, we need to spoof radio commands, prefereably
from the USB debug board.