Add battery over current warning
Added to the end of the board status enum, not in the more sensible area
of "near the other battery errors" because I don't want to screw up
which enum values map to what numbers.

I need this warning for radio board: It senses current both coming from
the battery and going out onto the bus. There's an error type for when
the bus is drawing too much current (E_BUS_OVER_CURRENT), but there
wasn't a way to report that too much current was being drawn from the
battery.

Just like the other current warnings, bytes 4 and 5 are for the
milliamperage that is being measured.