Add vent and injector battery voltages
For both VENT_VALVE_STAT and INJ_VALVE_STAT, bits 5 and 6 are the
battery voltages of the board's batteries (although the injector valve
does not currently have a different battery from the radio board (and
the rest of the can bus), it may in the future and thus it is just being
added now for simplicities sake.

The build_valve_stat_msg now has 2 more params for the 2 voltages that
are updated in the function, and the unit test was also updated
accordingly.