Ignore TXBP and IVRIF CAN interrupts on DSPIC
Logger is generating these errors and isn't handling them, which is
causing it to get stuck in an interrupt context. In the long term we
should figure out what is causing these and fix the root cause, but for
now I'm going to ignore it and focus on getting the logging part
working.