Fix compile issues on usbdbg with xc8v2.00
xc8v2 does a fun thing that v1.44 didn't, in which it detects a conflict
between the static variable `void (*spi_write)(...)` and the function
`void spi_write(...)` defined in cansw_usb.X/spi.c, and fails in the
link stage.

This commit fixes that compile error by prefixing the static function
pointers in mcp_2515.c with a double underscore, which seems like a good
idea since `spi_read/write` are probably common function names for
projects that use a spi based CAN transceiver. I thought that the C
standard guaranteed that names of static variables wouldn't clash with
global symbols in other object file, but apparently it doesn't.